Moved cache testing to the general tests
